Add the silence method
Addresses #37. This commit adds support for the `.silence()` protocol on
a logger, which retricts the logger from outputting messages below a
given level of severity.

Without support for this protocol, the logger may throw errors when used
in a Rails instance running in development mode. In dev mode, Rails can
be set to silence calls to log the loading of asset files.