build: resolves open-ended dependencies warning
When attempting to build the gem for publishing to the GitHub Package Repository, the CLI warned not to use an open-ended version, e.g. `>=0`, but use a bounded requirement, such as '~> x.y' for gem dependencies specified in the `.gemspec`. Also includes the updated CHANGELOG to reflect this as well as the incremented suffix for the version cadence.